app/databases/api/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.3
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.5.0
matplotlib-inline==0.1.7
multidict==6.0.5
mypy==1.10.0
mypy-extensions==1.0.0
nanoid==2.0.0
nest_asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==23.3.1
platformdirs==4.2.1
prompt-toolkit==3.0.42
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyzmq==25.1.2
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.18
stack-data==0.6.2
starlette==0.37.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
qdrant-client==1.9.1
Pillow
python-io
requests
fastapi
sentence-transformers==3.0.0

